workflow_router_agent/requirements.txt
# requirements.txt for workflow_router_agent
fastapi
uvicorn
httpx
python-dotenv
pydantic
